Refactor ProxyFactory and Safe
- Remove SafeBase and ProxyFactoryBase as they were breaking typing and IDE completion
- Refactor Safe creation logic into SafeCreator class
